goto-less entry symbol parent lookup
This goto was emulating the for-else structure from Python, which was required back when I was using a range-based for loop with std::ranges::reverse_view.  However, since I am currently forced to use iterators, I found another solution that does not use goto.
Personally, I was quite fond of this goto solution, but it is what it is.